[config/logLevel] provide basic logLevel setting
Ideally this is something we can flesh out a bit more down the road, but for now a simpley "logLevel" option is available at the top level of the cluster config. When set to one of the [standard log4j log levels](https://logging.apache.org/log4j/2.x/manual/customloglevels.html) that level will be used as the root log level, and the logging.yml file will be updated accordingly.

One caveat is that libesvm relies on the INFO logging of the node and http modules to learn where elasticsearch is listening, and when it is ready for requests. This means that for these two modules the log level can not be set below INFO (which the code enforces). At a later time we may want to support automatically changing the log levels to the requested one after we are done ready their INFO output (using the [transient cluster settings api](https://www.elastic.co/guide/en/elasticsearch/guide/current/logging.html) or something), but these modules don't seem very chatting at INFO level so I'm fine leaving it like this for now.